Move templates into the correct modules
