+++ src/csrs/pages/__init__.py
@@ -1,3 +1,3 @@
-from .home import Home
 from .loader import ENV
 from .static import errors
+from .templates.home import Home
