+++ src/csrs/pages/templates/__init__.py
@@ -6,10 +6,7 @@ from jinja2 import Environment, Template
 
 from ... import schemas
 from ..loader import ENV
-from .edit_assumption import EditableAssumption
-from .edit_path import EditablePaths
-from .edit_run import EditableRuns
-from .edit_scenario import EditableScenario
+from .editable import EditableAssumption, EditablePaths, EditableRuns, EditableScenario
 from .form_parts import (
     CreateSelection,
     CreateSelectionGroup,
@@ -17,7 +14,4 @@ from .form_parts import (
     CreateStrLong,
     CreateSwitch,
 )
-from .new_assumption import NewAssumption
-from .new_path import NewPath
-from .new_run import NewRuns
-from .new_scenario import NewScenario
+from .new import NewAssumption, NewPath, NewRuns, NewScenario
