+++ src/csrs/pages/templates/editable/__init__.py
@@ -1,0 +1,4 @@
+from .edit_assumption import EditableAssumption
+from .edit_path import EditablePaths
+from .edit_run import EditableRuns
+from .edit_scenario import EditableScenario
