+++ src/csrs/pages/templates/editable/edit_run.jinja
@@ -1,25 +1,25 @@
-{% extends 'templates/_editable.jinja' %}
+{% extends 'templates/editable/_editable.jinja' %}
 
 {% block title %}{{title|safe}}{% endblock %}
 {% block card %}
     <!-- Attributes  -->
     <div class="input-group input-group-sm m-0 p-0">
-        {{name|safe}}
+        {{scenario|safe}}
     </div>
     <div class="input-group input-group-sm my-1 p-0">
-        {{path|safe}}
+        {{version|safe}}
     </div>
     <div class="input-group input-group-sm my-1 pb-1">
-        {{category|safe}}
+        {{contact|safe}}
     </div>
     <div class="input-group input-group-sm my-1 pb-1">
-        {{period_type|safe}}
+        {{confidential|safe}}
     </div>
     <div class="input-group input-group-sm my-1 pb-1">
-        {{interval|safe}}
+        {{published|safe}}
     </div>
     <div class="input-group input-group-sm my-1 pb-1">
-        {{units|safe}}
+        {{code_version|safe}}
     </div>
     <div class="input-group input-group-sm my-1 pb-1">
         {{detail|safe}}
