+++ src/csrs/pages/templates/editable/edit_scenario.jinja
@@ -1,4 +1,4 @@
-{% extends 'templates/_editable.jinja' %}
+{% extends 'templates/editable/_editable.jinja' %}
 
 {% block title %}{{title}}{% endblock %}
 {% block card %}
