+++ src/csrs/pages/templates/home.py
@@ -3,14 +3,14 @@ from dataclasses import dataclass
 from fastapi import Request
 from jinja2 import Environment
 
-from .loader import ENV
+from ..loader import ENV
 
 
 @dataclass
 class Home:
     request: Request
     env: Environment = ENV
-    template: str = "home.jinja"
+    template: str = "templates/home.jinja"
 
     def __str__(self) -> str:
         return self.env.get_template(self.template).render(request=self.request)
