+++ src/csrs/pages/templates/new/__init__.py
@@ -1,0 +1,4 @@
+from .new_assumption import NewAssumption
+from .new_path import NewPath
+from .new_run import NewRuns
+from .new_scenario import NewScenario
