+++ src/csrs/pages/templates/new/new_path.jinja
@@ -1,4 +1,4 @@
-{% extends 'templates/_new.jinja' %}
+{% extends 'templates/new/_new.jinja' %}
 
 {% block title %}{{title}}{% endblock %}
 {% block card %}
