+++ src/csrs/pages/templates/new/new_run.py
@@ -1,9 +1,9 @@
 from fastapi import Request
 from jinja2 import Environment
 
-from ... import schemas
-from ..loader import ENV
-from .form_parts import CreateSelection, CreateStr, CreateStrLong, CreateSwitch
+from .... import schemas
+from ...loader import ENV
+from ..form_parts import CreateSelection, CreateStr, CreateStrLong, CreateSwitch
 
 
 class NewRuns:
@@ -65,7 +65,7 @@ class NewRuns:
         ).render(request)
 
         # render the whole card
-        return self.env.get_template("templates/new_run.jinja").render(
+        return self.env.get_template("templates/new/new_run.jinja").render(
             id=0,
             title="Create a New Run",
             scenario=scenario,
